requirements.txt
Django==1.6.1
PyPDF2==1.19
dj-database-url==0.2.2
dj-static==0.0.5
django-localflavor==1.0
gunicorn==18.0
pystache==0.5.3
django-vanilla-views==1.0.2
reportlab==2.7
static==1.0.2
wsgiref==0.1.2
